+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- HACK to fix the vm creation to update the disks attributes
+
 1.3.12
 ------